Add files via upload
I have created a code for reviweing and analyzing the comments that are posted on the comment section for a hotel named "Mejico" on "Yelp.com" website. By doing so using this code you can anazlize the sentiments of the comments and there by it will give an overall award to the hotel so you can choose the best hotel based on the comments given by the reviewers for that particular hotel or any product.

The code is reusable for other websites also and for reviewing other products too.
This way you can find out if the reviews given are genuine or not and choose the best product out there. 
Hope this helps. 
The Pytorch library is used in this to train model. The requests library is used to make HTTP requests. The BeautifulSoup library is used to parse HTML and XML documents. The re library is used for regular expressions. The transformers library is used to load and use pre-trained language models.
The nlptown/bert-base-multilingual-uncased-sentiment repository contains a pre-trained sentiment analysis model that is trained on a large corpus of text data in multiple languages.
The code uses the requests library to make an HTTP request to the Yelp website. The response is then parsed by the BeautifulSoup library to extract the reviews for Mejicano in Sydney.
Link of website:https://www.yelp.com/biz/mejico-sydney-2